+++ packages/hub/client-api/src/index.ts
@@ -1,6 +1,5 @@
 export * from './lib/guardConfigResponse';
 export * from './lib/hubService';
-export * from './lib/guardModel';
 export * from './lib/model';
 export * from './lib/serviceConfigResponse';
 export * from './lib/stanzaAuthTokenResponse';
