+++ packages/sdk-base/src/index.ts
@@ -2,6 +2,7 @@ export * from './lib/init';
 export * from './lib/initOrThrow';
 export * from './lib/guard/stanzaGuard';
 export * from './lib/guard/stanzaGuardError';
+export * from './lib/guard/model';
 export * from './lib/stanzaGuardHealth';
 export * from './lib/stanzaPriorityBoost';
 export * from './lib/utils/scheduler';
