+++ packages/sdk-base/src/lib/stanzaGuardHealth.ts
@@ -1,11 +1,6 @@
-import {
-  type Health,
-  type StanzaGuardHealthOptions,
-} from '@getstanza/hub-client-api';
 import { hubService } from './global/hubService';
+import { type StanzaGuardHealthOptions } from './guard/model';
 
-export const stanzaGuardHealth = async (
-  options: StanzaGuardHealthOptions
-): Promise<Health> => {
+export const stanzaGuardHealth = async (options: StanzaGuardHealthOptions) => {
   return hubService.getGuardHealth(options);
 };
